generator/requirements.txt
anthropic>=0.35.0
python-frontmatter>=1.0.0
pytz>=2024.1
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
